+++ src/main/strategies/MiddleTaxStrategy.java
@@ -1,7 +1,7 @@
 package main.strategies;
 
 public class MiddleTaxStrategy extends TaxStrategy {
-    public MiddleTaxStrategy(Double grossIncome) {
+    public MiddleTaxStrategy(long grossIncome) {
         super(grossIncome);
         this.setBasicTax(3572.0);
         this.setTaxPoint(37000.0);
